refactor: add PATCH and DELETE for Pedido

+++ app/Services/PedidoService.ts
@@ -28,13 +28,13 @@ class PedidoService {
       }
 
       if (estoque.quantidade < item.quantidade) {
-        throw new BadRequestException('Quantidade insuficiente em estoque', 400)
+        throw new BadRequestException(`Estoque do produto (id:${item.produtoId}) insuficiente`, 400)
       }
 
       return item
     }))
 
-    if(itensOfStock.length === 0) {
+    if (itensOfStock.length === 0) {
       throw new BadRequestException('Nenhum item foi adicionado ao pedido', 400)
     }
 
@@ -95,9 +95,102 @@ class PedidoService {
   }
 
   async delete(id: number) {
-    const pedido = await Pedido.findOrFail(id);
+    const pedido = await Pedido.findBy('id', id);
+
+    if (!pedido) {
+      throw new BadRequestException('Pedido não encontrado', 400)
+    }
+
+    const itens = await ItemPedido.query().where('pedidoId', pedido.id)
+
+    await Promise.all(itens.map(async (item) => {
+      const estoque = await Estoque.findBy('produtoId', item.produtoId)
+
+      if (!estoque) {
+        throw new BadRequestException(`Produto ${item.produtoId} não encontrado no estoque`, 400)
+      }
+
+      estoque.quantidade += item.quantidade
+      await item.delete()
+
+      await estoque.save()
+    }))
 
     await pedido.delete()
+  }
+
+  async update(id: number, data: PedidoCreateInput) {
+    const pedido = await Pedido.findBy('id', id);
+
+    if (!pedido) {
+      throw new BadRequestException('Pedido não encontrado', 400)
+    }
+
+    await pedido.load('itens')
+
+    if (data.clienteId) {
+      pedido.clienteId = data.clienteId
+    }
+
+    if (data.itens.length > 0) {
+      const itens = await ItemPedido.query().where('pedidoId', pedido.id)
+
+      await Promise.all(itens.map(async (item) => {
+        const estoque = await Estoque.findBy('produtoId', item.produtoId)
+
+        if (!estoque) {
+          throw new BadRequestException(`Produto ${item.produtoId} não encontrado no estoque`, 400)
+        }
+
+        estoque.quantidade += item.quantidade
+        await item.delete()
+
+        await estoque.save()
+      }))
+
+      let total = 0;
+
+      const itensOfStock = await Promise.all(data.itens.map(async (item) => {
+        const estoque = await Estoque.findBy('produtoId', item.produtoId)
+
+        if (!estoque) {
+          throw new BadRequestException(`Produto ${item.produtoId} não encontrado no estoque`, 400)
+        }
+
+        if (estoque.quantidade < item.quantidade) {
+          throw new BadRequestException(`Estoque do produto (id:${item.produtoId}) insuficiente`, 400)
+        }
+
+        return item
+      }))
+
+      await Promise.all(itensOfStock.map(async (item) => {
+        const estoque = await Estoque.findBy('produtoId', item.produtoId)
+        await estoque?.load('produto')
+
+        if (!estoque) {
+          throw new BadRequestException(`Produto ${item.produtoId} não encontrado no estoque`, 400)
+        }
+
+        estoque.quantidade -= item.quantidade
+        total += estoque.produto.preco * item.quantidade
+
+        await estoque.save()
+      }))
+
+      await ItemPedido.createMany(itensOfStock.map((item) => {
+        return {
+          produtoId: item.produtoId,
+          pedidoId: pedido.id,
+          quantidade: item.quantidade
+        }
+      }))
+
+      pedido.total = total
+      await pedido.save()
+    }
+
+    return { pedido };
   }
 }
